requirements.txt
fastapi
uvicorn
aiogram
sqlalchemy
ollama
python-dotenv
httpx
aiohttp
pydantic
pydantic-settings
jinja2
